backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
requests
aiohttp
aioftp
cryptography
acme
certbot
cloudflare
apscheduler
pytest
httpx
respx
boto3
godaddypy
configloader
bcrypt==3.2.0
passlib
python-jose[cryptography]
python-multipart
dnspython
sse-starlette
pydantic-settings
